feat: upload now has two options
