+++ frontend/src/app.pcss
@@ -119,5 +119,11 @@ button:focus:not(:focus-visible) {
 	white-space: nowrap;
 }
 
+.large-text {
+	font-size: 2.45rem;
+	font-weight: 500;
+	color: var(--primary-700);
+}
+
 @tailwind components;
 @tailwind utilities;
